requirements.txt
requests
beautifulsoup4
tqdm
Pillow
piexif
